feat : decoupled Authentication and names
Decoupled the Notion of UGS Players and Social profiles, to be reused across samples.
Created interfaces for authentication and moved the social profile interface with it.
The idea is to break this out into the plugin folder and reuse it in Parties, etc.
This fixes an issue with drag&drop where each of the standalone managers try to authenticate and throw errors.