Bugfix: Add FPS Limiter
Fixing a small bug - since there was no FPS limiter by default, the sample would use as much CPU as it could.  I've added a small fps limiting script and added it to some empty objects in the sample scenes named "FPS Limiter" to hopefully make it obvious to users.